pyproject.toml
[tool.poetry]
name = "home-page"
version = "0.1.0"
description = ""
authors = ["Julius Krahn <juliuskrahn@outlook.de>"]

[tool.poetry.dependencies]
python = "^3.9"
"aws-cdk.core" = "^1.128.0"
"aws-cdk.aws-cloudfront" = "^1.128.0"
"aws-cdk.aws-s3" = "^1.128.0"
"aws-cdk.aws-cloudfront-origins" = "^1.128.0"
"aws-cdk.aws-s3-deployment" = "^1.128.0"
"aws-cdk.aws-certificatemanager" = "^1.128.0"
"aws-cdk.aws-route53" = "^1.128.0"
"aws-cdk.aws-route53-targets" = "^1.128.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
